+++ src/vscode/treeviews/Workspaces/FabricGraphQLApis.ts
@@ -1,0 +1,44 @@
+import * as vscode from 'vscode';
+
+import { UniqueId } from '@utils/Helper';
+
+import { ThisExtension } from '../../../ThisExtension';
+import { iFabricApiItem } from '../../../fabric/_types';
+import { FabricApiService } from '../../../fabric/FabricApiService';
+import { FabricWorkspaceTreeItem } from './FabricWorkspaceTreeItem';
+import { FabricWorkspaceGenericFolder } from './FabricWorkspaceGenericFolder';
+import { FabricGraphQLApi } from './FabricGraphQLApi';
+
+// https://vshaxe.github.io/vscode-extern/vscode/TreeItem.html
+export class FabricGraphQLApis extends FabricWorkspaceGenericFolder {
+	constructor(
+		parent: FabricWorkspaceTreeItem
+	) {
+		super(`${parent.itemId}/GraphQL APIs`, "GraphQL APIs", "GraphQLApis", parent, "graphqlapis");
+
+		this.id = parent.itemId + "/" + parent.itemId + "/" + this.itemType.toString();
+	}
+
+	async getChildren(element?: FabricWorkspaceTreeItem): Promise<FabricWorkspaceTreeItem[]> {
+		if (element != null && element != undefined) {
+			return element.getChildren();
+		}
+		else {
+			let children: FabricGraphQLApi[] = [];
+
+			try {
+				const items = await FabricApiService.getList<iFabricApiItem>(this.apiPath);
+
+				for (let item of items.success) {
+					let treeItem = new FabricGraphQLApi(item, this);
+					children.push(treeItem);
+				}
+			}
+			catch (e) {
+				ThisExtension.Logger.logInfo("Could not load GraphQL APIs for workspace " + this.workspace.itemName);
+			}
+
+			return children;
+		}
+	}
+}