+++ src/vscode/treeviews/Workspaces/FabricWorkspace.ts
@@ -8,14 +8,14 @@ import { FabricLakehouses } from './FabricLakehouses';
 import { FabricApiItemType, FabricApiWorkspaceType, iFabricApiCapacity, iFabricApiItem, iFabricApiWorkspace } from '../../../fabric/_types';
 import { FabricApiService } from '../../../fabric/FabricApiService';
 import { FabricDataPipelines } from './FabricDataPipelines';
-import { FabricFSUri } from '../../filesystemProvider/FabricFSUri';
-import { FABRIC_SCHEME } from '../../filesystemProvider/FabricFileSystemProvider';
 import { FabricWorkspaceGenericFolder } from './FabricWorkspaceGenericFolder';
 import { FabricLakehouse } from './FabricLakehouse';
 import { FabricDataPipeline } from './FabricDataPipeline';
 import { FabricGitRepositories } from '../../sourceControl/FabricGitRepositories';
 import { FabricEnvironments } from './FabricEnvironments';
 import { FabricEnvironment } from './FabricEnvironment';
+import { FabricGraphQLApis } from './FabricGraphQLApis';
+import { FabricGraphQLApi } from './FabricGraphQLApi';
 
 // https://vshaxe.github.io/vscode-extern/vscode/TreeItem.html
 export class FabricWorkspace extends FabricWorkspaceTreeItem {
@@ -79,6 +79,9 @@ export class FabricWorkspace extends FabricWorkspaceTreeItem {
 						else if (item.type == "Environment") {
 							treeItem = new FabricEnvironments(this);
 						}
+						else if (item.type == "GraphQLApi") {
+							treeItem = new FabricGraphQLApis(this);
+						}
 						else {
 							treeItem = new FabricWorkspaceGenericFolder(
 								this.itemId + "/" + item.type + "s",
@@ -98,6 +101,9 @@ export class FabricWorkspace extends FabricWorkspaceTreeItem {
 					}
 					else if (item.type == "Environment") {
 						itemToAdd = new FabricEnvironment(item, this);
+					}
+					else if (item.type == "GraphQLApi") {
+						itemToAdd = new FabricGraphQLApi(item, this);
 					}
 					else {
 						itemToAdd = new FabricWorkspaceTreeItem(item.id, item.displayName, item.type as FabricApiItemType, itemTypes.get(item.type), item, item.description, vscode.TreeItemCollapsibleState.None);
@@ -119,7 +125,7 @@ export class FabricWorkspace extends FabricWorkspaceTreeItem {
 		// onelake:/<WorkspaceName>/<ItemName>.<ItemType>
 		const workspace = this.getParentByType<FabricWorkspace>("Workspace");
 
-		return vscode.Uri.parse(`onelake://${workspace.itemName}`);
+		return vscode.Uri.parse(`onelake://${workspace.itemId}`);
 	}
 
 	get apiUrlPart(): string {